+++ spec/rubocop/cop/rspec/describe_class_spec.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe RuboCop::Cop::RSpec::DescribeClass do
-  subject(:cop) { described_class.new }
-
+RSpec.describe RuboCop::Cop::RSpec::DescribeClass, :config do
   it 'checks first-line describe statements' do
     expect_offense(<<-RUBY)
       describe "bad describe" do
@@ -120,57 +118,6 @@ RSpec.describe RuboCop::Cop::RSpec::DescribeClass do
     end
   end
 
-  it 'ignores request specs' do
-    expect_no_offenses(<<-RUBY)
-      describe 'my new feature', type: :request do
-      end
-    RUBY
-  end
-
-  it 'ignores feature specs' do
-    expect_no_offenses(<<-RUBY)
-      describe 'my new feature', type: :feature do
-      end
-    RUBY
-  end
-
-  it 'ignores system specs' do
-    expect_no_offenses(<<-RUBY)
-      describe 'my new system test', type: :system do
-      end
-    RUBY
-  end
-
-  it 'ignores feature specs when RSpec.describe is used' do
-    expect_no_offenses(<<-RUBY)
-      RSpec.describe 'my new feature', type: :feature do
-      end
-    RUBY
-  end
-
-  it 'flags specs with non :type metadata' do
-    expect_offense(<<-RUBY)
-      describe 'my new feature', foo: :feature do
-               ^^^^^^^^^^^^^^^^ The first argument to describe should be the class or module being tested.
-      end
-    RUBY
-  end
-
-  it 'flags normal metadata in describe' do
-    expect_offense(<<-RUBY)
-      describe 'my new feature', blah, type: :wow do
-               ^^^^^^^^^^^^^^^^ The first argument to describe should be the class or module being tested.
-      end
-    RUBY
-  end
-
-  it 'ignores feature specs - also with complex options' do
-    expect_no_offenses(<<-RUBY)
-      describe 'my new feature', :test, :type => :feature, :foo => :bar do
-      end
-    RUBY
-  end
-
   it 'ignores an empty describe' do
     expect_no_offenses(<<-RUBY)
       RSpec.describe do
@@ -179,24 +126,6 @@ RSpec.describe RuboCop::Cop::RSpec::DescribeClass do
       describe do
       end
     RUBY
-  end
-
-  it 'ignores routing specs' do
-    expect_no_offenses(<<-RUBY)
-      describe 'my new route', type: :routing do
-      end
-    RUBY
-  end
-
-  it 'ignores view specs' do
-    expect_no_offenses(<<-RUBY)
-      describe 'widgets/index', type: :view do
-      end
-    RUBY
-  end
-
-  it "doesn't blow up on single-line describes" do
-    expect_no_offenses('describe Some::Class')
   end
 
   it "doesn't flag top level describe in a shared example" do
@@ -233,5 +162,57 @@ RSpec.describe RuboCop::Cop::RSpec::DescribeClass do
         end
       end
     RUBY
+  end
+
+  it 'ignores `type` metadata ignored by default' do
+    expect_no_offenses(<<-RUBY)
+      describe 'widgets/index', type: :view do
+      end
+    RUBY
+  end
+
+  it 'flags specs with non `type` metadata' do
+    expect_offense(<<-RUBY)
+      describe 'foo bar', foo: :bar do
+               ^^^^^^^^^ The first argument to describe should be the class or module being tested.
+      end
+    RUBY
+  end
+
+  it 'ignores feature specs - also with complex options' do
+    expect_no_offenses(<<-RUBY)
+      describe 'my new feature', :test, foo: :bar, type: :feature do
+      end
+    RUBY
+  end
+
+  it 'flags non-ignored `type` metadata' do
+    expect_offense(<<-RUBY)
+      describe 'wow', blah, type: :wow do
+               ^^^^^ The first argument to describe should be the class or module being tested.
+      end
+    RUBY
+  end
+
+  context 'when IgnoredMetadata is configured' do
+    let(:cop_config) do
+      { 'IgnoredMetadata' =>
+        { 'foo' => ['bar'],
+          'type' => ['wow'] } }
+    end
+
+    it 'ignores configured metadata' do
+      expect_no_offenses(<<-RUBY)
+        describe 'foo bar', foo: :bar do
+        end
+      RUBY
+    end
+
+    it 'ignores configured `type` metadata' do
+      expect_no_offenses(<<-RUBY)
+        describe 'my new system test', type: :wow do
+        end
+      RUBY
+    end
   end
 end